requirements.txt
requests==2.31.0
dnspython==2.4.2
flask==3.0.0
